src/web/bot/requirements.txt
requests
python-telegram-bot
sqlalchemy
psycopg2
schedule
graphyte
appmetrics